+++ frontend/app/src/components/App.svelte
@@ -7,14 +7,21 @@
     import { rtlStore } from "../stores/rtl";
     import { _, isLoading } from "svelte-i18n";
     import Router from "./Router.svelte";
-    import { notFound, pathParams } from "../routes";
+    import { notFound, pathParams, routeForScope } from "../routes";
     import SwitchDomain from "./SwitchDomain.svelte";
     import Upgrading from "./upgrading/Upgrading.svelte";
     import UpgradeBanner from "./UpgradeBanner.svelte";
     import { currentTheme } from "../theme/themes";
     import "../stores/fontSize";
     import Profiler from "./Profiler.svelte";
-    import { OpenChat, UserLoggedIn, type DiamondMembershipFees } from "openchat-client";
+    import {
+        OpenChat,
+        UserLoggedIn,
+        type DiamondMembershipFees,
+        type ChatSummary,
+        type ChatIdentifier,
+        routeForChatIdentifier,
+    } from "openchat-client";
     import { type UpdateMarketMakerConfigArgs, inititaliseLogger } from "openchat-client";
     import {
         isCanisterUrl,
@@ -32,6 +39,9 @@
     import Head from "./Head.svelte";
     import { snowing } from "../stores/snow";
     import Snow from "./Snow.svelte";
+    import ActiveCall from "./home/video/ActiveCall.svelte";
+    import { incomingVideoCall } from "../stores/video";
+    import IncomingCall from "./home/video/IncomingCall.svelte";
     overrideItemIdKeyNameBeforeInitialisingDndZones("_id");
 
     const logger = inititaliseLogger(
@@ -71,9 +81,11 @@
     let client: OpenChat = createOpenChatClient();
 
     let profileTrace = client.showTrace();
+    let videoCallElement: ActiveCall;
 
     setContext<OpenChat>("client", client);
 
+    $: chatListScope = client.chatListScope;
     $: identityState = client.identityState;
     $: landingPageRoute = isLandingPageRoute($pathParams);
     $: anonUser = client.anonUser;
@@ -354,6 +366,19 @@
         calculateHeight();
     }
 
+    function startVideoCall(ev: CustomEvent<{ chat: ChatSummary; join: boolean }>) {
+        videoCallElement?.startOrJoinVideoCall(ev.detail.chat, ev.detail.join);
+    }
+
+    function joinVideoCall(ev: CustomEvent<ChatIdentifier>) {
+        incomingVideoCall.set(undefined);
+        const chat = client.lookupChatSummary(ev.detail);
+        if (chat) {
+            page(routeForChatIdentifier("none", chat.id));
+            videoCallElement?.startOrJoinVideoCall(chat, true);
+        }
+    }
+
     let isFirefox = navigator.userAgent.indexOf("Firefox") >= 0;
     $: burstPath = $currentTheme.mode === "dark" ? "/assets/burst_dark" : "/assets/burst_light";
     $: burstUrl = isFirefox ? `${burstPath}.png` : `${burstPath}.svg`;
@@ -369,6 +394,12 @@
 
 <Head />
 
+<ActiveCall
+    on:clearSelection={() => page(routeForScope($chatListScope))}
+    bind:this={videoCallElement} />
+
+<IncomingCall on:joinVideoCall={joinVideoCall} />
+
 <Witch background />
 
 {#if isCanisterUrl}
@@ -377,7 +408,7 @@
     <Upgrading />
 {:else if $identityState.kind === "anon" || $identityState.kind === "logging_in" || $identityState.kind === "registering" || $identityState.kind === "logged_in" || $identityState.kind === "loading_user"}
     {#if !$isLoading}
-        <Router {showLandingPage} />
+        <Router on:startVideoCall={startVideoCall} {showLandingPage} />
     {/if}
 {/if}
