+++ frontend/app/src/components/home/HomeRoute.svelte
@@ -24,7 +24,7 @@
         </div>
     </div>
 {:else}
-    <Home />
+    <Home on:startVideoCall />
 {/if}
 
 <style lang="scss">
